Basic functionality complete
The code is not very pretty, but I was in a rush. Supports
two run modes: timestamped (-tai64) input for filling in
historical data, and live input for consuming the stdout of
dnscache directly. I tried to focus on making sure it exits
when input is exhausted or output has a problem, so that
it can be supervised properly.