Starting work on a utility to find slow tx queries
The dnscache-slowlogs utility reads a dnscache log file
and looks for slow queries to remote servers. Currently
it is not sufficient; I am struggling with matching the
outgoing requests ("tx" lines) with their responses ("rr"
lines). So you will get a ton of false positive with slow
transactions with ridiculous time differences (1hr etc).

I may have misunderstood the log file format for dnscache,
and so will have to go back to the drawing board on this one.
Putting it up here for posterity.